move: Const Enum분리
